Change main background loop and misc changes

+++ cogs/server.py
@@ -3,7 +3,11 @@ import re
 
 from discord.ext import commands
 
+from modules import utils
+from modules.embeds import admin_notice
+
 minecraft_dir = '/home/artilapx/minecraft/'
+
 
 #  BadCoder thanks
 
@@ -21,41 +25,48 @@ def attach_session():
     os.system("tmux attach -t minecraft")
 
 
-class Server(commands.Cog, command_attrs=dict(hidden=True), name="Server"):
-    """Команды для взаимодействия с сервером Minecraft. Мы не можем их показать в целях безопасности."""
+class Server(commands.Cog, name="Server"):
+    """Команды для взаимодействия с сервером Minecraft."""
 
     def __init__(self, bot):
         self.bot = bot
 
     @commands.command()
-    @commands.has_permissions(administrator=True)
     async def start(self, ctx):
         """Запускает сервер Rumblur."""
-        if not session_exists():
-            os.system("mtc start --only")
-            await ctx.send(f"`Отправлен запрос на включение сервера...`")
+        if utils.is_admin(ctx.message.author.id):
+            if not session_exists():
+                os.system("mtc start --only")
+                await ctx.send(f"`Отправлен запрос на включение сервера...`")
+            else:
+                await ctx.send(f"`Сервер уже запущен и работает. Запуск не требуется.`")
         else:
-            await ctx.send(f"`Сервер уже запущен и работает. Запуск не требуется.`")
+            await ctx.send(embed=await admin_notice())
 
     @commands.command()
-    @commands.has_permissions(administrator=True)
     async def restart(self, ctx):
         """Перезагружает сервер Rumblur."""
-        if session_exists():
-            os.system("mtc restart --now")
-            await ctx.send(f"`Отправлен запрос на перезагрузку сервера...`")
+        if utils.is_admin(ctx.message.author.id):
+            if session_exists():
+                os.system("mtc restart --now")
+                await ctx.send(f"`Отправлен запрос на перезагрузку сервера...`")
+            else:
+                await ctx.send(f"`Сервер выключен. Нечего перезагружать.`")
         else:
-            await ctx.send(f"`Сервер выключен. Нечего перезагружать.`")
+            await ctx.send(embed=await admin_notice())
 
-    @commands.command()
-    @commands.has_permissions(administrator=True)
-    async def stop(self, ctx):
-        """Останавливает сервер Rumblur."""
+
+@commands.command()
+async def stop(self, ctx):
+    """Останавливает сервер Rumblur."""
+    if utils.is_admin(ctx.message.author.id):
         if session_exists():
             os.system("mtc shutdown --confirm")
             await ctx.send(f"`Отправлен запрос на выключение сервера...`")
         else:
             await ctx.send(f"`Сервер уже выключен. Остановка не требуется.`")
+    else:
+        await ctx.send(embed=await admin_notice())
 
 
 def setup(bot):
